+++ app/src/main/java/com/example/gestiontienda2/data/local/entities/entity/OrderEntity.kt
@@ -5,9 +5,9 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "orders")
 data class OrderEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val orderId: Int, // Assuming this is a foreign key to an Order entity
-    val clientId: Int,
+    val clientId: Long,
     val orderDate: Long, // Using Long for timestamp
     val status: String,
     val totalAmount: Double,
